server/requirements.txt
fastapi
openai
python-dotenv